Zero'ing out the frazil and the surface fluxes below the icebergs
A piece of code has been added which zeros out the surface fluxes and the ocean frazil underneath very large icebergs. This code is mostly copied from the ice shelf module.
The latent heat, sensible, SW, LW, evap and lprec have all been set to zero. The frazil is set to zero, and the appropriate quantity is added to evap.